+++ CHANGELOG.md
@@ -109,6 +109,7 @@ Updates:
 Removed:
 
 - `NA#modules.child_process` removed.
+- `NA#modules.forcedomain` removed.
 - `NA#webconfig.componentsRelativePath` removed.
 - `NA#locals.urlBasePathSlice` removed.
 - `NA#variations.currentRouteName` removed.